Factiva_News/requirements.txt
# Core Data Science and ML Libraries
numpy>=1.21.0
pandas>=1.3.0
#scikit-learn>=1.0.0
scipy>=1.7.0
pydantic>=2.0.0

# NLP and Topic Modeling
#bertopic
#sentence-transformers>=2.2.0
gensim>=4.2.0
#octis>=1.10.0
transformers>=4.20.0
torch>=1.12.0
tiktoken>=0.4.0

# Clustering and Dimensionality Reduction
umap-learn>=0.5.3
hdbscan>=0.8.28

# OpenAI API
openai>=1.0.0

# Instructor library for structured outputs
instructor>=1.0.0

# Optional provider-specific dependencies
# Google Generative AI
google-generativeai>=0.3.0

# Anthropic
anthropic>=0.20.0

# Web Scraping and Data Collection
newspaper3k>=0.2.8
feedparser>=6.0.8
requests>=2.28.0
beautifulsoup4>=4.11.0
urllib3>=1.26.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Utility Libraries
tqdm>=4.64.0
tenacity>=8.0.0
python-dateutil>=2.8.0
pytz>=2022.1

# Data Storage and Processing
pickle-mixin>=1.0.2
json5>=0.9.6

# Development and Testing
pytest>=7.0.0
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# Optional: Environment Management
python-dotenv>=0.19.0

# Optional: Advanced NLP
spacy>=3.4.0
nltk>=3.7.0 